Mark `eval_opt_chain` as pure
