+++ ecmascript/minifier/src/compress/util.rs
@@ -1,4 +1,5 @@
 use crate::debug::dump;
+use swc_atoms::js_word;
 use swc_common::DUMMY_SP;
 use swc_ecma_ast::*;
 use swc_ecma_transforms_base::ext::MapWithMut;
@@ -194,6 +195,31 @@ pub(crate) fn is_ok_to_negate_rhs(rhs: &Expr) -> bool {
             false
         }
     }
+}
+
+pub(crate) fn is_pure_undefined(e: &Expr) -> bool {
+    match e {
+        Expr::Ident(Ident {
+            sym: js_word!("undefined"),
+            ..
+        }) => true,
+
+        Expr::Unary(UnaryExpr {
+            op: UnaryOp::Void,
+            arg,
+            ..
+        }) if !arg.may_have_side_effects() => true,
+
+        _ => false,
+    }
+}
+
+pub(crate) fn is_pure_undefined_or_null(e: &Expr) -> bool {
+    is_pure_undefined(e)
+        || match e {
+            Expr::Lit(Lit::Null(..)) => true,
+            _ => false,
+        }
 }
 
 /// A negative value means that it's efficient to negate the expression.
